+++ sympy/core/tests/test_args.py
@@ -2988,8 +2988,8 @@ def test_sympy__matrices__expressions__matexpr__MatrixSymbol():
     assert _test_args(MatrixSymbol('A', 3, 5))
 
 
-def test_sympy__matrices__expressions__matexpr__OneMatrix():
-    from sympy.matrices.expressions.matexpr import OneMatrix
+def test_sympy__matrices__expressions__special__OneMatrix():
+    from sympy.matrices.expressions.special import OneMatrix
     assert _test_args(OneMatrix(3, 5))
