+++ sympy/matrices/expressions/__init__.py
@@ -6,7 +6,7 @@ from .companion import CompanionMatrix
 from .funcmatrix import FunctionMatrix
 from .inverse import Inverse
 from .matadd import MatAdd
-from .matexpr import MatrixExpr, MatrixSymbol, OneMatrix, matrix_symbols
+from .matexpr import MatrixExpr, MatrixSymbol, matrix_symbols
 from .matmul import MatMul
 from .matpow import MatPow
 from .trace import Trace, trace
@@ -19,7 +19,7 @@ from .dotproduct import DotProduct
 from .kronecker import kronecker_product, KroneckerProduct, combine_kronecker
 from .permutation import PermutationMatrix, MatrixPermute
 from .sets import MatrixSet
-from .special import ZeroMatrix, Identity
+from .special import ZeroMatrix, Identity, OneMatrix
 
 __all__ = [
     'MatrixSlice',
