+++ sympy/matrices/expressions/tests/test_determinant.py
@@ -4,7 +4,7 @@ from sympy.matrices.expressions import (
     Identity, MatrixExpr, MatrixSymbol, Determinant,
     det, ZeroMatrix, Transpose
 )
-from sympy.matrices.expressions.matexpr import OneMatrix
+from sympy.matrices.expressions.special import OneMatrix
 from sympy.testing.pytest import raises
 from sympy import refine, Q
