+++ sympy/matrices/expressions/tests/test_matexpr.py
@@ -1,13 +1,14 @@
-from sympy import (KroneckerDelta, diff, Piecewise, Sum, Dummy, factor,
+from sympy import (KroneckerDelta, diff, Sum, Dummy, factor,
                    expand, zeros, gcd_terms, Eq, Symbol)
 
 from sympy.core import S, symbols, Add, Mul, SympifyError, Rational
-from sympy.functions import transpose, sin, cos, sqrt, cbrt, exp
+from sympy.functions import sin, cos, sqrt, cbrt, exp
 from sympy.simplify import simplify
-from sympy.matrices import (Identity, ImmutableMatrix, Inverse, MatAdd, MatMul,
-        MatPow, Matrix, MatrixExpr, MatrixSymbol, ShapeError, ZeroMatrix,
+from sympy.matrices import (ImmutableMatrix, Inverse, MatAdd, MatMul,
+        MatPow, Matrix, MatrixExpr, MatrixSymbol, ShapeError,
         SparseMatrix, Transpose, Adjoint, NonSquareMatrixError, MatrixSet)
-from sympy.matrices.expressions.matexpr import MatrixElement, OneMatrix
+from sympy.matrices.expressions.matexpr import MatrixElement
+from sympy.matrices.expressions.special import ZeroMatrix, Identity
 from sympy.testing.pytest import raises, XFAIL
 
 
@@ -39,25 +40,6 @@ def test_matrix_symbol_creation():
     raises(ValueError, lambda: MatrixSymbol('A', n, n))
 
 
-def test_one_matrix_creation():
-    assert OneMatrix(2, 2)
-    assert OneMatrix(0, 0)
-    assert Eq(OneMatrix(1, 1), Identity(1))
-    raises(ValueError, lambda: OneMatrix(-1, 2))
-    raises(ValueError, lambda: OneMatrix(2.0, 2))
-    raises(ValueError, lambda: OneMatrix(2j, 2))
-    raises(ValueError, lambda: OneMatrix(2, -1))
-    raises(ValueError, lambda: OneMatrix(2, 2.0))
-    raises(ValueError, lambda: OneMatrix(2, 2j))
-
-    n = symbols('n')
-    assert OneMatrix(n, n)
-    n = symbols('n', integer=False)
-    raises(ValueError, lambda: OneMatrix(n, n))
-    n = symbols('n', negative=True)
-    raises(ValueError, lambda: OneMatrix(n, n))
-
-
 def test_shape():
     assert A.shape == (n, m)
     assert (A*B).shape == (n, l)
@@ -85,123 +67,6 @@ def test_subs():
     C, D = MatrixSymbol('C', 2, 2), MatrixSymbol('D', 2, 2)
 
     assert (C*D).subs({C: A, D: B}) == MatMul(A, B)
-
-
-def test_ZeroMatrix():
-    A = MatrixSymbol('A', n, m)
-    Z = ZeroMatrix(n, m)
-
-    assert A + Z == A
-    assert A*Z.T == ZeroMatrix(n, n)
-    assert Z*A.T == ZeroMatrix(n, n)
-    assert A - A == ZeroMatrix(*A.shape)
-
-    assert Z
-
-    assert transpose(Z) == ZeroMatrix(m, n)
-    assert Z.conjugate() == Z
-
-    assert ZeroMatrix(n, n)**0 == Identity(n)
-    with raises(NonSquareMatrixError):
-        Z**0
-    with raises(NonSquareMatrixError):
-        Z**1
-    with raises(NonSquareMatrixError):
-        Z**2
-
-    assert ZeroMatrix(3, 3).as_explicit() == ImmutableMatrix.zeros(3, 3)
-
-
-def test_ZeroMatrix_doit():
-    Znn = ZeroMatrix(Add(n, n, evaluate=False), n)
-    assert isinstance(Znn.rows, Add)
-    assert Znn.doit() == ZeroMatrix(2*n, n)
-    assert isinstance(Znn.doit().rows, Mul)
-
-
-def test_OneMatrix():
-    A = MatrixSymbol('A', n, m)
-    a = MatrixSymbol('a', n, 1)
-    U = OneMatrix(n, m)
-
-    assert U.shape == (n, m)
-    assert isinstance(A + U, Add)
-    assert transpose(U) == OneMatrix(m, n)
-    assert U.conjugate() == U
-
-    assert OneMatrix(n, n) ** 0 == Identity(n)
-    with raises(NonSquareMatrixError):
-        U ** 0
-    with raises(NonSquareMatrixError):
-        U ** 1
-    with raises(NonSquareMatrixError):
-        U ** 2
-    with raises(ShapeError):
-        a + U
-
-    U = OneMatrix(n, n)
-    assert U[1, 2] == 1
-
-    U = OneMatrix(2, 3)
-    assert U.as_explicit() == ImmutableMatrix.ones(2, 3)
-
-
-def test_OneMatrix_doit():
-    Unn = OneMatrix(Add(n, n, evaluate=False), n)
-    assert isinstance(Unn.rows, Add)
-    assert Unn.doit() == OneMatrix(2 * n, n)
-    assert isinstance(Unn.doit().rows, Mul)
-
-
-def test_OneMatrix_mul():
-    assert OneMatrix(n, m) * OneMatrix(m, k) == OneMatrix(n, k) * m
-    assert w * OneMatrix(1, 1) == w
-    assert OneMatrix(1, 1) * w.T == w.T
-
-
-def test_Identity():
-    A = MatrixSymbol('A', n, m)
-    i, j = symbols('i j')
-
-    In = Identity(n)
-    Im = Identity(m)
-
-    assert A*Im == A
-    assert In*A == A
-
-    assert transpose(In) == In
-    assert In.inverse() == In
-    assert In.conjugate() == In
-
-    assert In[i, j] != 0
-    assert Sum(In[i, j], (i, 0, n-1), (j, 0, n-1)).subs(n,3).doit() == 3
-    assert Sum(Sum(In[i, j], (i, 0, n-1)), (j, 0, n-1)).subs(n,3).doit() == 3
-
-    # If range exceeds the limit `(0, n-1)`, do not remove `Piecewise`:
-    expr = Sum(In[i, j], (i, 0, n-1))
-    assert expr.doit() == 1
-    expr = Sum(In[i, j], (i, 0, n-2))
-    assert expr.doit().dummy_eq(
-        Piecewise(
-            (1, (j >= 0) & (j <= n-2)),
-            (0, True)
-        )
-    )
-    expr = Sum(In[i, j], (i, 1, n-1))
-    assert expr.doit().dummy_eq(
-        Piecewise(
-            (1, (j >= 1) & (j <= n-1)),
-            (0, True)
-        )
-    )
-    assert Identity(3).as_explicit() == ImmutableMatrix.eye(3)
-
-
-def test_Identity_doit():
-    Inn = Identity(Add(n, n, evaluate=False))
-    assert isinstance(Inn.rows, Add)
-    assert Inn.doit() == Identity(2*n)
-    assert isinstance(Inn.doit().rows, Mul)
 
 
 def test_addition():
